# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for IAD (Information and Analytical Data) system.",
        "title": "IAD Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/your-org/iad-backend",
            "email": "support@iad.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/stage-request-stages/{requestId}/stages/{stageId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update stage connection details in a stage request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-request-stages"
                ],
                "summary": "Update stage in request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage Request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Stage ID",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stage connection update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateStageToRequestConnection"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a stage from a stage request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stage-request-stages"
                ],
                "summary": "Remove stage from request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Stage Request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Stage ID",
                        "name": "stageId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.UpdateStageToRequestConnection": {
            "type": "object",
            "properties": {
                "input_field_1": {
                    "type": "integer"
                },
                "input_field_2": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}